Move cargo commands to separate file
