+++ src/cargo.rs
@@ -1,0 +1,64 @@
+use std::process::{exit, Command};
+
+pub struct ArgBuilder(Vec<String>);
+
+impl ArgBuilder {
+    pub fn new() -> Self {
+        Self(Vec::new())
+    }
+
+    /// Add an argument.
+    ///
+    /// Can be a `&str` or a `String`.
+    /// If it contains whitespace, it is split into multiple args.
+    pub fn add<A>(&mut self, arg: A)
+    where
+        A: Into<String>,
+    {
+        let arg: String = arg.into();
+
+        for part in arg.split_ascii_whitespace() {
+            self.0.push(part.to_string());
+        }
+    }
+
+    /// Add an argument with a value.
+    pub fn add_with_value<A, V>(&mut self, arg: A, value: V)
+    where
+        A: Into<String>,
+        V: Into<String>,
+    {
+        self.add(arg);
+        self.add(value);
+    }
+}
+
+/// Run `cargo run` with the given arguments.
+///
+/// Exits the program if `cargo run` failed.
+pub fn cargo_run(args: ArgBuilder) {
+    let status = Command::new("cargo")
+        .arg("run")
+        .args(args.0)
+        .status()
+        .expect("Failed to run `cargo run`.");
+
+    if !status.success() {
+        exit(status.code().unwrap_or(1));
+    }
+}
+
+/// Run `cargo build` with the given arguments.
+///
+/// Exits the program if `cargo run` failed.
+pub fn cargo_build(args: ArgBuilder) {
+    let status = Command::new("cargo")
+        .arg("build")
+        .args(args.0)
+        .status()
+        .expect("Failed to run `cargo build`.");
+
+    if !status.success() {
+        exit(status.code().unwrap_or(1));
+    }
+}
