+++ src/main.rs
@@ -1,6 +1,7 @@
 use clap::Parser;
 use cli::{CargoCommand, Command};
 
+pub(crate) mod cargo;
 mod cli;
 mod new;
 mod run;
